requirements.txt
aiogram~=3.4.1
telebot~=0.0.5
python-dotenv~=1.0.1